requirements.txt
streamlit>=1.54.0
python-dotenv>=1.2.1
st-copy-to-clipboard>=0.1.6
requests>=2.32.5
beautifulsoup4>=4.14.3
youtube_transcript_api>=1.2.4
google-generativeai>=0.8.6
lxml>=6.0.2
urllib3>=2.6.3
yt-dlp>=2024.0.0